+++ deltann/core/io.h
@@ -21,6 +21,7 @@ limitations under the License.
 
 #include "core/misc.h"
 #include "core/shape.h"
+#include "core/buffer.h"
 
 #ifdef USE_TF
 #include "tensorflow/core/framework/tensor_shape.h"
@@ -38,8 +39,8 @@ enum class InOut { DELTA_IN = 0, DELTA_OUT };
 // model input/output metadata
 class BaseInOut {
  public:
-  BaseInOut(std::string name, int id, Shape shape, DataType dtype)
-      : _name(name), _id(id), _shape(shape), _dtype(dtype) {}
+  BaseInOut(std::string name, int id, Shape shape, DataType dtype, InOut which)
+      : _name(name), _id(id), _shape(shape), _dtype(dtype), _inout(which) {}
 
   virtual ~BaseInOut() {}
 
@@ -89,15 +90,11 @@ class BaseInOut {
 class Input : public BaseInOut {
  public:
   Input(std::string name, int id, Shape shape, DataType dtype)
-      : BaseInOut(name, id, shape, dtype) {
-    _inout = InOut::DELTA_IN;
+      : BaseInOut(name, id, shape, dtype, InOut::DELTA_IN) {
   }
-
   Input(std::string name, int id, DataType dtype)
-      : BaseInOut(name, id, Shape(), dtype) {
-    _inout = InOut::DELTA_IN;
+      : BaseInOut(name, id, Shape(), dtype, InOut::DELTA_IN) {
   }
-
   ~Input() {}
 };
 
@@ -105,17 +102,294 @@ class Input : public BaseInOut {
 class Output : public BaseInOut {
  public:
   Output(std::string name, int id, Shape shape, DataType dtype)
-      : BaseInOut(name, id, shape, dtype) {
-    _inout = InOut::DELTA_OUT;
+      : BaseInOut(name, id, shape, dtype, InOut::DELTA_OUT) {
   }
-
   Output(std::string name, int id, DataType dtype)
-      : BaseInOut(name, id, Shape(), dtype) {
-    _inout = InOut::DELTA_OUT;
+      : BaseInOut(name, id, Shape(), dtype, InOut::DELTA_OUT) {
+  }
+  ~Output() {}
+
+
+
+};
+
+class BaseInOutData {
+ public:
+  explicit BaseInOutData(BaseInOut& inout) : _inout(inout) {
+    if (inout.dtype() != DataType::DELTA_NONE && 0 != inout.size()) {
+      _data = std::make_shared<Buffer>(inout.size() *
+                                       delta_dtype_size(inout.dtype()));
+    } else {
+      _data = std::make_shared<Buffer>();
+    }
   }
 
-  ~Output() {}
+  ~BaseInOutData() { _data = nullptr; }
+
+  void* ptr() const { return _data->ptr(); }
+
+  int ndim(void) const { return _inout.ndim(); }
+
+  Shape& shape(void) { return _inout.shape(); }
+  const Shape& shape(void) const { return _inout.shape(); }
+
+  const size_t size(void) const { return _inout.size(); }
+
+  const size_t bytes(void) const {
+    return _inout.size() * delta_dtype_size(_inout.dtype());
+  }
+
+  DataType dtype() const { return _inout.dtype(); }
+
+  const std::string name() const { return _inout.name(); }
+
+  int id() const { return _inout.id(); }
+
+  void set_shape(const Shape& shape) { _inout.set_shape(shape); }
+
+#ifdef USE_TF
+  void set_shape(const tensorflow::TensorShape& shape) {
+    _inout.set_shape(shape);
+  }
+#endif
+
+  void set_dtype(const DataType dtype) { _inout.set_dtype(dtype); }
+
+  void feed_random_data() {
+    DataType dtype = this->dtype();
+    switch (dtype) {
+      case DataType::DELTA_FLOAT32:
+        make_random(static_cast<float*>(_data->ptr()), size());
+        break;
+      case DataType::DELTA_INT32:
+        std::fill_n(static_cast<int*>(_data->ptr()), size(), 1);
+        break;
+      case DataType::DELTA_CHAR:
+        for (int i = 0; i < size(); ++i) {
+          char* ptr = static_cast<char*>(_data->ptr());
+          std::fill_n(ptr + i, 1, gen_random_char());
+        }
+        break;
+      default:
+        LOG_FATAL << "not support datatype";
+    }
+  }
+
+  void copy_from(const void* src, std::size_t size) {
+    DataType dtype = this->dtype();
+    if (dtype != DataType::DELTA_NONE) {
+      std::size_t bytes = size * delta_dtype_size(dtype);
+      resize(bytes);
+      _data->copy_from(src, bytes);
+    } else {
+      LOG_FATAL << "_dtype is DataType::DELTA_NONE ";
+    }
+  }
+
+  void copy_from(const float* src) { copy_from(src, this->size()); }
+
+  void copy_from(const float* src, std::size_t size) {
+    DataType dtype = this->dtype();
+    if (dtype != DataType::DELTA_NONE) {
+      _data->copy_from(src, size * delta_dtype_size(dtype));
+    } else {
+      LOG_FATAL << "_dtype is DataType::DELTA_NONE ";
+    }
+  }
+
+  void copy_to(void* dst, std::size_t byte_size) {
+    _data->copy_to(dst, byte_size);
+  }
+
+  void resize(const std::size_t size) {
+    DataType dtype = this->dtype();
+    if (dtype != DataType::DELTA_NONE) {
+      _data->resize(size * delta_dtype_size(dtype));
+    } else {
+      LOG_FATAL << "Data type is DELTA_NONE";
+    }
+  }
+
+#ifdef USE_TF
+  tensorflow::TensorShape tensor_shape() const {
+    tensorflow::Status status;
+    const Shape& shape = this->shape();
+    int ndim = this->ndim();
+    std::vector<int> tmp_shape;
+    for (int i = 0; i < ndim; ++i) {
+      tmp_shape.push_back(shape[i]);
+    }
+    tensorflow::TensorShape ts;
+    status = tensorflow::TensorShapeUtils::MakeShape(tmp_shape, &ts);
+    if (!status.ok()) {
+      LOG_FATAL << "Error when make shape from vector: " << tmp_shape << ", "
+                << status;
+    }
+    return ts;
+  }
+#endif
+
+  virtual BaseInOut& meta(void) const { return _inout; }
+
+ protected:
+  BaseInOut& _inout;
+  std::shared_ptr<Buffer> _data;
 };
+
+class InputData : public BaseInOutData {
+ public:
+  explicit InputData(BaseInOut& inout) : BaseInOutData(inout) {}
+};
+
+class OutputData : public BaseInOutData {
+ public:
+  explicit OutputData(BaseInOut& inout) : BaseInOutData(inout) {}
+};
+
+
+
+// class BaseInOutData {
+//  public:
+//   explicit BaseInOutData(BaseInOut& inout) : _inout(inout) {
+//     allocate();
+//   }
+// 
+//   void allocate() {
+//     if (dtype() != DataType::DELTA_NONE && 0 != size()) {
+//       _data = std::make_shared<Buffer>(bytes());
+//     } else {
+//       _data = std::make_shared<Buffer>();
+//     }
+//   }
+// 
+//   ~BaseInOutData() { _data = nullptr; }
+// 
+//   void* ptr() const { return _data->ptr(); }
+//   
+//   // elements
+//   const size_t size(void) const { return _inout.size(); }
+// 
+//   DataType dtype() const { return _inout.dtype(); }
+// 
+//   void set_dtype(const DataType dtype) { _inout.set_dtype(dtype); }
+// 
+//   const size_t bytes(void) const {
+//     return size() * delta_dtype_size(dtype());
+//   }
+// 
+//   const std::string name() const { return _inout.name(); }
+// 
+//   int id() const { return _inout.id(); }
+// 
+//   Shape& shape(void) { return _inout.shape(); }
+// 
+//   Shape& shape(void) const { return _inout.shape(); }
+// 
+//   int ndim(void) const { return _inout.ndim(); }
+// 
+// #ifdef USE_TF
+//   void set_shape(const tensorflow::TensorShape& shape) {
+//     _inout.set_shape(shape);
+//   }
+// #endif
+//   void set_shape(const Shape& shape) { _inout.set_shape(shape); }
+// 
+//   void feed_random_data() {
+//     DataType type = dtype();
+//     switch (type) {
+//       case DataType::DELTA_FLOAT32:
+//         make_random(static_cast<float*>(_data->ptr()), size());
+//         break;
+//       case DataType::DELTA_INT32:
+//         std::fill_n(static_cast<int*>(_data->ptr()), size(), 1);
+//         break;
+//       case DataType::DELTA_CHAR:
+//         for (int i = 0; i < size(); ++i) {
+//           char* ptr = static_cast<char*>(_data->ptr());
+//           std::fill_n(ptr + i, 1, gen_random_char());
+//         }
+//         break;
+//       default:
+//         LOG_FATAL << "not support datatype";
+//     }
+//   }
+// 
+//   void copy_from(const void* src, std::size_t nelems) {
+//     DataType type = this->dtype();
+//     if (type != DataType::DELTA_NONE) {
+//       std::size_t bytes = nelems * delta_dtype_size(type);
+//       resize(bytes);
+//       _data->copy_from(src, bytes);
+//     } else {
+//       LOG_FATAL << "_dtype is DataType::DELTA_NONE ";
+//     }
+//   }
+// 
+//   void copy_from(const float* src) { copy_from(src, this->size()); }
+// 
+//   void copy_from(const float* src, std::size_t size) {
+//     // copy_from(static_cast<const void*>(src), size);
+//     DataType dtype = this->dtype();
+//     if (dtype != DataType::DELTA_NONE) {
+//       _data->copy_from(src, size * delta_dtype_size(dtype));
+//     } else {
+//       LOG_FATAL << "_dtype is DataType::DELTA_NONE ";
+//     }
+//   }
+// 
+//   void copy_to(void* dst, std::size_t byte_size) {
+//     _data->copy_to(dst, byte_size);
+//   }
+// 
+//   void resize(const std::size_t size) {
+//     DataType dtype = this->dtype();
+//     if (dtype != DataType::DELTA_NONE) {
+//        if (_data->size() != size) _data->resize(size);
+//     } else {
+//       LOG_FATAL << "Data type is DELTA_NONE";
+//     }
+//   }
+// 
+// #ifdef USE_TF
+//   tensorflow::TensorShape tensor_shape() const {
+//     tensorflow::Status status;
+//     Shape& s = shape();
+//     //int ndim = this->ndim();
+// 
+//     // std::vector<int> tmp_shape;
+//     // for (int i = 0; i < ndim; ++i) {
+//     //   tmp_shape.push_back(s[i]);
+//     // }
+//     std::vector<int> tmp_shape = s.vec();
+// 
+//     tensorflow::TensorShape ts;
+//     status = tensorflow::TensorShapeUtils::MakeShape(tmp_shape, &ts);
+//     if (!status.ok()) {
+//       LOG_FATAL << "Error when make shape from vector: " << tmp_shape << ", "
+//                 << status;
+//     }
+//     return ts;
+//   }
+// #endif
+// 
+//   virtual BaseInOut& meta(void) const { return _inout; }
+// 
+//  protected:
+//   BaseInOut& _inout;
+//   std::shared_ptr<Buffer> _data;
+// };
+// 
+// class InputData : public BaseInOutData {
+//  public:
+//   explicit InputData(BaseInOut& inout) : BaseInOutData(inout) {}
+// };
+// 
+// class OutputData : public BaseInOutData {
+//  public:
+//   explicit OutputData(BaseInOut& inout) : BaseInOutData(inout) {}
+// };
+
+
 
 }  // namespace core
